requirements.txt
flask
flask[async]
requests
python-dotenv
flask_socketio
httpx
Pillow
eventlet
discord.py